+++ src/constants/index.js
@@ -1,0 +1,3 @@
+export const BASE_SERVER_URL =
+  "https://react-shoppingcart-server.herokuapp.com";
+export const PRODUCT_LIST_PATH = "/products";
